requirements.txt
fastapi
uvicorn
google-generativeai
pypdf2
python-multipart
python-dotenv